swarm: rename netsim package to network/simulation
