+++ swarm/network/simulation/kademlia_test.go
@@ -13,7 +13,8 @@
 //
 // You should have received a copy of the GNU Lesser General Public License
 // along with the go-ethereum library. If not, see <http://www.gnu.org/licenses/>.
-package netsim
+
+package simulation
 
 import (
 	"context"
